==> Code/0. extract pdf text.ipynb <==
import fitz  # PyMuPDF
import os

def extract_text_from_pdf(pdf_path, txt_path):
    # Open the PDF file
    pdf_document = fitz.open(pdf_path)
    text = ""
    
    # Iterate over each page in the PDF
    for page_num in range(len(pdf_document)):
        # Extract text from the page
        page = pdf_document.load_page(page_num)
        text += page.get_text()
    
    # Write the text to a .txt file
    with open(txt_path, 'w', encoding='utf-8') as txt_file:
        txt_file.write(text)
        
    print(f"Text has been successfully extracted to {txt_path}")

# PDF files Directory
pdf_directory = 'Respiratory_System'

# output
output_directory = 'Respiratory_System'
os.makedirs(output_directory, exist_ok=True)

# Loop through all the files in the directory
for filename in os.listdir(pdf_directory):
    if filename.endswith('.pdf'):
        pdf_path = os.path.join(pdf_directory, filename)
        txt_filename = os.path.splitext(filename)[0] + '.txt'
        txt_path = os.path.join(output_directory, txt_filename)
        
        extract_text_from_pdf(pdf_path, txt_path)


import nltk
from nltk.corpus import stopwords
from nltk.probability import FreqDist
from nltk.collocations import BigramCollocationFinder
from nltk.metrics import BigramAssocMeasures

txt_path = 'Oxygen.txt'

# with open(txt_path, 'r') as file:
    # text = file.read()
    
# need encode to read specific char
with open(txt_path, 'r', encoding='utf-8') as file:
    text = file.read()
    
tokens = nltk.word_tokenize(text)

stop_words = set(stopwords.words('english'))
tokens = [word for word in tokens if word.isalpha() and word.lower() not in stop_words]

#freq_dist = FreqDist.tokenize(tokens) #notworking 
freq_dist = FreqDist(tokens)

for word, frequency in freq_dist.most_common(20):
    print(f"{word}: {frequency}")

bigram_finder = BigramCollocationFinder.from_words(tokens)

bigrams = bigram_finder.nbest(BigramAssocMeasures.likelihood_ratio, 20)

print("\nTop 20 bigrams:")
for bigram in bigrams:
    print(bigram)
