==> Code/4. extract keyword.ipynb <==
import pandas as pd
import os
from unidecode import unidecode

# Initialize lists to store the titles, paths, and document contents
titles = []
paths = []
docs = []

# Define the directory you want to traverse
directory = 'Dataset'

# Traverse the directory
for root, dirs, files in os.walk(directory):
    for filename in files:
        if filename.endswith('.txt'):  # For each text file
            titles.append(os.path.splitext(filename)[0])
            txt_filename = os.path.splitext(filename)[0] + '.txt'
            txt_path = os.path.join(root, txt_filename)
            
            paths.append(txt_path)
            #print(txt_path)
            with open(txt_path, encoding="utf8") as fo:
                docs.append(unidecode(fo.read()))  # Append the text to docs


from tqdm import tqdm
from keybert import KeyBERT

keywords = []
kw_model = KeyBERT()

for doc in tqdm(docs, desc="Extracting Keywords"):
    keywords.append(kw_model.extract_keywords(doc, stop_words='english'))

print(keywords[0][0])

data=[]
for d in range(len(docs)):
    # print(d, len(keywords[d]))
    if(len(keywords[d]) > 0):
        kw=keywords[d]
        data.append([titles[d], docs[d], kw[0][0], kw[0][1], kw[1][0],kw[1][1],kw[2][0],kw[2][1], kw[3][0],kw[3][1], kw[4][0],kw[4][1]])

#print(data[0][1])
#print(docs[21])
df=pd.DataFrame(data=data, columns=["Title","Text", "Keyword1","Score1", "Keyword2", "Score2", "Keyword3", "Score3", "Keyword4","Score4","Keyword5", "Score5"])

df.head()
df.to_csv('keyword.csv')
